FIX: 사용 중인 카테고리 삭제 차단
카테고리 삭제 시 활성 챌린지/스토어가 존재하면 CATEGORY_IN_USE(409)
예외를 던지도록 변경. CategoryMapper에 countActive 조회 메서드와
SQL을 추가하고, ExceptionCode에 신규 코드 등록.

+++ src/main/java/sisosolsol/greenfire/category/service/CategoryService.java
@@ -8,6 +8,8 @@ import sisosolsol.greenfire.category.model.dto.CategoryCreateDTO;
 import sisosolsol.greenfire.category.model.dto.CategoryDTO;
 import sisosolsol.greenfire.category.model.dto.CategoryUpdateDTO;
 import sisosolsol.greenfire.common.enums.category.CategoryType;
+import sisosolsol.greenfire.common.exception.CustomException;
+import sisosolsol.greenfire.common.exception.type.ExceptionCode;
 
 import java.util.List;
 
@@ -54,13 +56,18 @@ public class CategoryService {
     }
 
     public void deleteCategory(Integer categoryCode, CategoryType categoryType) {
-        // TODO: 각 카테고리에 등록된 챌린지/스토어가 없을 때에만 삭제 허용
         int result = 0;
         switch (categoryType) {
             case CHALLENGE:
+                if (categoryMapper.countActiveChallengesByCategoryCode(categoryCode) > 0) {
+                    throw new CustomException(ExceptionCode.CATEGORY_IN_USE);
+                }
                 result = categoryMapper.deleteChallengeCategory(categoryCode);
                 break;
             case STORE:
+                if (categoryMapper.countActiveStoresByCategoryCode(categoryCode) > 0) {
+                    throw new CustomException(ExceptionCode.CATEGORY_IN_USE);
+                }
                 result = categoryMapper.deleteStoreCategory(categoryCode);
                 break;
         }
